% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-chebyshev.r
\name{cheb_1st}
\alias{cheb_1st}
\title{Generating Chebyshev Polynomials of first kind}
\usage{
cheb_1st(x_axis, n)
}
\arguments{
\item{x_axis}{Arbitrary x-axis}

\item{n}{Order of the polynomial}
}
\value{
Chebyshev polynomial of the first kind
}
\description{
\code{cheb_1st} generates Chebyshev polynomials of the first kind of the 
  order n from an arbitrary x-axis. Recurrence formula is taken from 
  Bronstein p.xxx.
}
