% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-chebyshev.r
\name{cheb_2nd}
\alias{cheb_2nd}
\title{Generating Chebyshev Polynomials of second kind}
\usage{
cheb_2nd(x_axis, n)
}
\arguments{
\item{x_axis}{Arbitrary x-axis}

\item{n}{Order of the polynomial}
}
\value{
Chebyshev polynomials of the second kind
}
\description{
\code{cheb_2nd} generates Chebyshev polynomials of the second kind of the 
  order n from an arbitrary x-axis. Recurrence formula is taken from 
  Bronstein p.xxx.
}
