% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-chebyshev.r
\name{cheb_deriv_1st}
\alias{cheb_deriv_1st}
\title{Calculation of the values of the first derivation}
\usage{
cheb_deriv_1st(x_axis, cheb_coeff)
}
\arguments{
\item{x_axis}{Arbitrary x-axis}

\item{cheb_coeff}{Chebyshev-coefficients of the least squares model fit}
}
\value{
Values of the first derivation of the filtered 
  model
}
\description{
\code{cheb_deriv_1st} calculates values of the first derivation of the 
  filtered model from the Chebyshev coefficients of the least squares fit on 
  the original x-axis or single x-values.
}
