% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-chebyshev.r
\name{cheb_find_extr}
\alias{cheb_find_extr}
\title{Curve Fitting with Chebyshev Polynomials and Finding of its Roots}
\usage{
cheb_find_extr(d, x_axis, n, roots_bound_lower = NA, roots_bound_upper = NA)
}
\arguments{
\item{d}{Data series to be fitted}

\item{x_axis}{Arbitrary x-axis}

\item{n}{Order of the polynomial}

\item{roots_bound_lower}{Lower boundary of the extreme value search}

\item{roots_bound_upper}{Upper boundary of the extreme value search}
}
\value{
list of calculated parameters like Chebyshev coefficients,
  values of the filtered model, values of the first and seconda derivation,
  postitions and values of extremes
}
\description{
\code{cheb_fit_extr} fits a Chebyshev polynomial of order n to a data series 
  using the method of least squares and locates the extrema of the fit using 
  the Newton-Raphson-algorithm.
}
