% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-chebyshev.r
\name{cheb_find_max}
\alias{cheb_find_max}
\title{Routine to find maximum values of a curve}
\usage{
cheb_find_max(d, x_axis, n, maxima_bound_lower = NA,
  maxima_bound_upper = NA)
}
\arguments{
\item{d}{Data series to be fitted}

\item{x_axis}{Arbitrary x-axis}

\item{n}{Order of the polynomial}

\item{maxima_bound_lower}{Lower boundary of the maximum value search}

\item{maxima_bound_upper}{Upper boundary of the maximum value search}
}
\value{
list of calculated parameters like Chebyshev coefficients,
  values of the filtered model, values of the first and seconda derivation,
  postitions and values of maxima
}
\description{
\code{cheb_find_max} fits a Chebyshev polynomial of order n to a data series 
  using the method of least squares, locates the extrema of the fit using 
  the Newton-Raphson-algorithm of the first derivation and checks for maxima 
  using the second derivative.
}
