% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-chebyshev.r
\name{cheb_fit}
\alias{cheb_fit}
\title{Curve Fitting with Chebyshev Polynomials}
\usage{
cheb_fit(d, x_axis, n)
}
\arguments{
\item{d}{Data series to be fitted}

\item{x_axis}{Arbitrary X-Axis}

\item{n}{Order of the polynomial}
}
\value{
Values of the filtered model
}
\description{
\code{cheb_fit} fits a Chebyshev polynomial of order n to a arbitrary data 
  or time series using the method of least squares.
}
