% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-chebyshev.r
\name{cheb_model_filter}
\alias{cheb_model_filter}
\title{Calculation of Values of the model fit}
\usage{
cheb_model_filter(x_axis, cheb_coeff)
}
\arguments{
\item{x_axis}{Arbitrary x-axis or single value}

\item{cheb_coeff}{Coefficients of the Least-Squares-Fit with Chebyshev
polynomials}
}
\value{
Filtered model (or single value)
}
\description{
\code{cheb_model} calculates filtered model values of arbitrary
}
