% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-chebyshev.r
\name{cheb_rescale}
\alias{cheb_rescale}
\title{Rescaling of X-Axis to Original X-Axis}
\usage{
cheb_rescale(x_cheb, x_axis)
}
\arguments{
\item{x_cheb}{Scaled x-axis}

\item{x_axis}{Unscaled x-axis}
}
\value{
Rescaled x-axis
}
\description{
\code{cheb_rescale} rescales (single or multiple) values in the range [-1,1] 
to the original x-axis.
}
