% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-chebyshev.r
\name{cheb_scale}
\alias{cheb_scale}
\title{Scaling of X-Axis to [-1,1]}
\usage{
cheb_scale(x_axis, x_val = NA)
}
\arguments{
\item{x_axis}{Arbitrary x-axis}

\item{x_val}{Single value in range of x_axis}
}
\value{
Depending on input parameters: Either scaled x-axis or scaled scalar 
  value
}
\description{
\code{cheb_scale} scales an arbitrary x-axis to the range [-1, 1], so it can 
  be used for a Chebyshev polynomial, which is defined on this span.
}
